Drop class inheritance
* Use Protocol
* Drop class inheritance
* Starting to clean up both Apple and Slack provider
